Cypress - merge tests and refactor mock use / login
